refactor: разделить классы по пакетам
