+++ src/main/java/ru/practicum/shareit/booking/dto/BookingDto.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import ru.practicum.shareit.booking.BookingStatus;
+import ru.practicum.shareit.booking.model.BookingStatus;
 
 import javax.validation.constraints.FutureOrPresent;
 import javax.validation.constraints.Positive;
