+++ src/main/java/ru/practicum/shareit/user/UserMapper.java
@@ -1,5 +1,8 @@
 package ru.practicum.shareit.user;
 
+import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.model.User;
+
 public class UserMapper {
     public static UserDto toUserDto(User user) {
         return new UserDto(user.getId(), user.getName(),
