+++ src/main/java/ru/practicum/shareit/user/model/User.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.model;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -13,11 +13,20 @@ import javax.validation.constraints.NotNull;
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class UserDto {
+public class User implements Cloneable {
     private Long id;
     @NotBlank
     private String name;
     @NotNull
     @Email
     private String email;
+
+    @Override
+    public User clone() {
+        try {
+            return (User) super.clone();
+        } catch (CloneNotSupportedException e) {
+            throw new AssertionError();
+        }
+    }
 }
