+++ src/main/java/ru/practicum/shareit/user/storage/UserStorage.java
@@ -1,4 +1,7 @@
-package ru.practicum.shareit.user;
+package ru.practicum.shareit.user.storage;
+
+import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.dto.UserDto;
 
 import java.util.Collection;
 import java.util.Optional;
