+++ src/main/java/ru/practicum/shareit/utils/ErrorHandler.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import ru.practicum.shareit.utils.exceptions.*;
+import ru.practicum.shareit.utils.model.ErrorResponse;
 
 @Slf4j
 @RestControllerAdvice
