+++ src/main/java/ru/practicum/shareit/utils/model/ErrorResponse.java
@@ -1,4 +1,4 @@
-package ru.practicum.shareit.utils;
+package ru.practicum.shareit.utils.model;
 
 public class ErrorResponse {
     private final String error;
