+++ src/main/java/ru/practicum/shareit/utils/service/ValidationService.java
@@ -1,11 +1,11 @@
-package ru.practicum.shareit.utils;
+package ru.practicum.shareit.utils.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
-import ru.practicum.shareit.user.User;
-import ru.practicum.shareit.user.UserDto;
-import ru.practicum.shareit.user.UserStorage;
+import ru.practicum.shareit.user.model.User;
+import ru.practicum.shareit.user.dto.UserDto;
+import ru.practicum.shareit.user.storage.UserStorage;
 import ru.practicum.shareit.utils.exceptions.NotUniqueValueException;
 
 import javax.validation.Valid;
